CASS-37 | Create CassandraBaseOperation to extract execution builder.
